Smart App / Download buttons
Buttons serve as user funnels.

Funnel nr 1: Umbrel, get ItchySats on the Appstore
Funnel nr 2: Download the binary (system-specific download link based on latest release)

The buttons work in the way, that there is a default link that points to getumbrel and the latest release page.
On page ready we use JQuery to check for locally running Umbrel and latest release by operating system and switch the links if we are able to construct them properly.